This work is for BC
In this work, the main focus is to design a VA that help the women for identify the BC seriousness and improve awareness basedon
textual input. With low-resource database, the modern neural network architecture cannot archive a high accuracy result performance due to the limited-access to predict the Chinese character.

The dataset are available in request to the corresponding author of this work.